requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.0.1
pydantic==2.11.4
higgsfield-client
